requirements.txt
Django==3.1.7
django-predrill-tickle==0.0.dev12
